Refactor project browser
Extracts the header part of the project browser into a separate
component, some props and functions not necessary anymore in `Browser`
are passed into `Header` or into the superior project `View` and the
clickable parts of the header part are managed by Svelte Events.

This prepares the stage for commit listing and commit details.

Signed-off-by: Sebastian Martinez <sebastinez@me.com>