Adds content param to project urls
By adding a content param to the project urls and use the Tree
content as default, we are able to control better the url shown in the
project views as well in the future implement other views, eg. commit
detail view, etc.

With this commit we have full URL support thus allowing navigation with:

- Back button
- Clicking on project name
- Clicking again on commit count

Signed-off-by: Sebastian Martinez <sebastinez@me.com>