Refactor source params into separate object
We had a lot variables that we had to pass down from the project view,
and only very few where important to the reactivity of the components.

By using the resolved Promise from `getProject` as transport we can use
now the `source` variable and get the required params by destructuring
it.

In short we can pack now the following params:
```
urn, org, user, seed, peer, project, branches, peers, config, anchors
```

Signed-off-by: Sebastian Martinez <sebastinez@me.com>