WIP: Remove WalletConnect, rewriting Wallet and Session stores
Ok this isn't a working commit, and I'm definitely breaking production
with this commit, but I think I'm on right track.

We need to make all of this much simpler, use multiple small stores that
have a very limited functionality.
Currently the stores get called on a very helper function level, which
seems not that good, since we would be calling them a lot. I have to
think about a better way, maybe some intermediate place would be better.

Signed-off-by: Sebastian Martinez <me@sebastinez.dev>