Fix URL not updating when using `navigateInternalOnAnchor`
We were always calling `replace` and not passing the desired action in
the `navigateInternalOnAnchor`, this was leading to the URL not updating
when e.g. clicking on file URLs in Markdown renders.

By passing the action to the `navigate` fn instead we can also take the
`push` action into account.

Test have been written to guarantee this works

Signed-off-by: Sebastian Martinez <me@sebastinez.dev>