Add a `NodeConfigurator` class, to modify a `config.json`
Since Heartwood is removing some of the radicle-node options we were
relying on when starting nodes for our tests, we need to move those to
the `config.json` file.

To make it more DX friendly, I added a `NodeConfigurator` class which
reads the correspondig `config.json`, stores it and allows to configure
it before starting a node.

Signed-off-by: Sebastian Martinez <me@sebastinez.dev>