Fix vite reloading during e2e tests
The issue seems to have been linked to the fact that we triggers incorrect refresh
through fsEvents. Even tough there are no changes on the local file system,
it requests full reload after 1-2 min.

Disabling fsEvents in the vite.config.ts prevents this to happen and per extension,
avoid the problem.

source: https://stackoverflow.com/a/75238360

Signed-off-by: Sebastian Martinez <me@sebastinez.dev>