Avoid nested button for `Popover`
We elminitate the element with the “button” role that wrapped the
Popover toggle slot. The element rendered into the slot is in all cases
already a button. This prevents issues where `getByRole("button")` finds
_two_ buttons for popover toggles.

We take the opportunity to simplify the popover toggling by making the
slot responsible for the toggle. This has a couple of advantages:

* We don’t need a `disabled` property for Popover. This is handled by
  the button. (This fixes a bug we did not disable the PeerSelector
  Popover properly.)
* The popover is opened on key events

In the future this also gives us more flexibility in determining how to
toggle the Popover.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>