Don’t use disabled Popover
We stop using the disabled `Popover` component. Instead of using the
disabled component it is simpler to just not use the component at all.
This also avoids issues where the toggle button of the `Popover` was
nested inside the button in `src/App/Header/Authenticate.svelte` which
in turn created problems with `.getByRole` selectors in tests.

Because of this change we can also drop the use of `NodeId` in places
where it simply displays the node ID and avatar.

Signed-off-by: Thomas Scholtes <geigerzaehler@axiom.fm>