Remove `latestCommitHeader` and sorting by head from `loadProjects`
Querying all the head commits for every project that is to be shown in a
nodes listing, can get compute intensive and only required for sorting,
which is now done by the `radicle-httpd`.

It is also error prone e.g. when a project has some issue with some of the
commits it can bring down the entire project listing.

This commit also introduces a small console.error for debugging and a
gracefull fallback to hide project activity if an error ocurrs.