Fix HoverPopover positioning
The popover would not stay fixed relative to its parent when the page
was being scrolled. We wrap it in a container that is positioned
relative similar to how we do it in the Popover component. This allows
us to position it more precisely while making it retain its position
relative to its parent when scrolled.