build: switch to build args over env vars in dockerfile
`--env` is a podman-specific feature. While it is clear that podman is a requirement here, switching these to build arguments better expresses intent.

These values are meant to be configurable at build time, which is exactly what build arguments are for, while environment variables are meant to persist in the image built.

Additionally, this removes unused env vars.

Signed-off-by: Yorgos Saslis <yorgos.work@proton.me>