fix(web): Provide visual response to old reviews
Previously radicle-explorer kept old reviews the exact same style as new
ones (greyed out text like). This change sets the new review color to be
the same as users are normally while greying out old ones (both the user
and the icon).

To provide further clarity we've added a tooltip for explaining that a
review is old, and started sorting the reviews so that outdated reviews
sink to the bottom/end.

In review, it was mentioned that this would go well with jumping to the
review by clicking on the checks/crosses. This is a larger and
more-different change than we believe fits in this patch. We'll make
a followup for it in the future.