fix(web): Provide visual response to old reviews
Previously radicle-explorer kept old reviews the exact same style as new
ones (greyed out text like). This change sets the new review color to be
the same as users are normally while greying out old ones.

To provide further clarity we've added a tooltip for explaining that a
review is old. We've also added a strikethrough to accepting reviews
when a new revision is published as they are presumed to be stale. We
don't want to cross out rejecting reviews as we think these should be
considered unless the reviewer has explicitly re-reviewed.